% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NeuralNetTools_utils.R
\name{bias_lines}
\alias{bias_lines}
\title{Plot connection weights for bias lines}
\usage{
bias_lines(bias_x, bias_y, mod_in, nid, rel_rsc, all_out, pos_col, neg_col,
  struct, y_names, x_range, y_range, layer_x, line_stag, max_sp)
}
\arguments{
\item{bias_x}{numeric vector x axis locations for bias lines}

\item{bias_y}{numeric vector y axis locations for bias lines}

\item{mod_in}{neural network model object}

\item{nid}{logical value indicating if neural interpretation diagram is plotted, default \code{TRUE}}

\item{rel_rsc}{numeric indicating the scaling range for the width of connection weights in a neural interpretation diagram. Default is \code{NULL} for no rescaling.}

\item{all_out}{chr string indicating names of response variables for which connections are plotted, default all}

\item{pos_col}{chr string indicating color of positive connection weights, default \code{'black'}}

\item{neg_col}{chr string indicating color of negative connection weights, default \code{'grey'}}

\item{struct}{numeric vector for network structure}

\item{y_names}{chr string for names of output variables}

\item{x_range}{numeric of x axis range for base plot}

\item{y_range}{numeric of x axis range for base plot}

\item{layer_x}{numeric indicating locations of layers on x axis}

\item{line_stag}{numeric value that specifies distance of connection weights from nodes}

\item{max_sp}{logical indicating if space is maximized in plot}
}
\description{
Plot connection weights for bias lines in \code{\link{plotnet}}
}

