% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetTools_utils.R
\name{bias_points}
\alias{bias_points}
\title{Plot bias points}
\usage{
bias_points(
  bias_x,
  bias_y,
  layer_name,
  node_labs,
  x_range,
  y_range,
  circle_cex,
  cex_val,
  bord_col,
  circle_col
)
}
\arguments{
\item{bias_x}{numeric vector of values for x locations}

\item{bias_y}{numeric vector for y location}

\item{layer_name}{string indicating text to put in node}

\item{node_labs}{logical indicating of node labels are included}

\item{x_range}{numeric of x axis range for base plot}

\item{y_range}{numeric of y axis range for base plot}

\item{circle_cex}{numeric value indicating size of nodes, default 5}

\item{cex_val}{numeric value indicating size of text labels, default 1}

\item{bord_col}{chr string indicating border color around nodes, default \code{'lightblue'}}

\item{circle_col}{chr string indicating color of nodes}
}
\description{
Plot bias points in \code{\link{plotnet}}
}
