% Generated by roxygen2 (4.0.2): do not edit by hand
\name{garson}
\alias{garson}
\title{Modification of Garson's algorithm for variable importance}
\usage{
garson(out.var, mod.in, bar.plot = T, struct = NULL, x.lab = NULL,
  y.lab = NULL, wts.only = F)
}
\arguments{
\item{out.var}{}

\item{mod.in}{}

\item{bar.plot}{}

\item{struct}{}

\item{x.lab}{}

\item{y.lab}{}

\item{wts.only}{}
}
\value{
what it returns
}
\description{
Modification of Garson's algorithm for variable importance
}

