% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetTools_utils.R
\name{get_ys}
\alias{get_ys}
\title{Get y locations for layers in \code{\link{plotnet}}}
\usage{
get_ys(lyr, max_sp, struct, y_range)
}
\arguments{
\item{lyr}{numeric indicating layer for getting y locations}

\item{max_sp}{logical indicating if space is maximized in plot}

\item{struct}{numeric vector for network structure}

\item{y_range}{numeric vector indicating limits of y axis}
}
\description{
Get y locations for input, hidden, output layers in \code{\link{plotnet}}
}

