% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetTools_neuraldat.R
\docType{data}
\name{neuraldat}
\alias{neuraldat}
\title{Simulated dataset for function examples}
\format{
A data frame with 2000 rows and 5 variables:
\describe{
  \item{Y1}{numeric}
  \item{Y2}{numeric}
  \item{X1}{numeric}
  \item{X2}{numeric}
  \item{X3}{numeric}
  ...
}
}
\usage{
neuraldat
}
\description{
A simulated dataset of 2000 observations containing two response variables and three explanatory variables.  Explanatory variables were sampled from a standard normal distribution.  Response variables were linear combinations of the explanatory variables.  The response variables Y1 and Y2 are standardized from 0 to 1.
}
\keyword{datasets}
