% Generated by roxygen2 (4.0.2): do not edit by hand
\name{neuralweights}
\alias{neuralweights}
\title{Get weights for a neural network}
\usage{
neuralweights(mod.in, nid, rel.rsc, struct.out)
}
\arguments{
\item{mod.in}{}

\item{nid}{}

\item{rel.rsc}{}

\item{struct.out}{}
}
\value{
Returns \code{list} of weight values for the input model
}
\description{
Get weights for a neural network
}

