% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetTools_utils.R
\name{lekgrps}
\alias{lekgrps}
\title{Create optional barplot for \code{\link{lekprofile}} groups}
\usage{
lekgrps(grps, position = "dodge", grp_nms = NULL)
}
\arguments{
\item{grps}{\code{\link[base]{data.frame}} of values for each variable in each group used to create groups in \code{\link{lekprofile}}}

\item{position}{chr string indicating bar position (e.g., 'dodge', 'fill', 'stack'), passed to \code{\link[ggplot2]{geom_bar}}}

\item{grp_nms}{optional chr string of alternative names for groups in legend}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Create optional barplot of constant values of each variable for each group used with \code{\link{lekprofile}}
}
\examples{
## enters used with kmeans clustering
x <- neuraldat[, c('X1', 'X2', 'X3')]
grps <- kmeans(x, 6)$center

lekgrps(grps)
}

