% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lekprofile}
\alias{lekprofile}
\title{Sensitivity analysis based on Lek's profile method}
\usage{
lekprofile(mod.in, var.sens = NULL, resp.name = NULL, exp.in = NULL,
  steps = 100, split.vals = seq(0, 1, by = 0.2), val.out = F)
}
\arguments{
\item{mod.in}{}

\item{var.sens}{}

\item{resp.name}{}

\item{exp.in}{}

\item{steps}{}

\item{split.vals}{}

\item{val.out}{}
}
\value{
what it returns
}
\description{
Sensitivity analysis based on Lek's profile method
}

