% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.nnet}
\alias{plot.nnet}
\title{Plot a neural network model}
\usage{
\method{plot}{nnet}(mod.in, nid = T, all.out = T, all.in = T, bias = T,
  wts.only = F, rel.rsc = 5, circle.cex = 5, node.labs = T,
  var.labs = T, x.lab = NULL, y.lab = NULL, line.stag = NULL,
  struct = NULL, cex.val = 1, alpha.val = 1, circle.col = "lightblue",
  pos.col = "black", neg.col = "grey", bord.col = "lightblue",
  max.sp = F, ...)
}
\arguments{
\item{mod.in}{}

\item{nid}{}

\item{all.out}{}

\item{all.in}{}

\item{bias}{}

\item{wts.only}{}

\item{rel.rsc}{}

\item{circle.cex}{}

\item{node.labs}{}

\item{var.labs}{}

\item{x.lab}{}

\item{y.lab}{}

\item{line.stag}{}

\item{struct}{}

\item{cex.val}{}

\item{alpha.val}{}

\item{circle.col}{}

\item{pos.col}{}

\item{neg.col}{}

\item{max.sp}{}

\item{...}{}

\item{bor.col}{}
}
\value{
what -it returns
}
\description{
Plot a neural network model
}
\examples{
rnorm(10)
}

